Create `HackrNewsFeedStoreSpecs` that must be imoplemented by any 'Store' test implementation to guarantee it meets specs
