+++ HackrNewsTests/Stories Cache/CodableHackrNewsFeedStoreTests.swift
@@ -5,33 +5,10 @@
 import HackrNews
 import XCTest
 
-protocol HackrNewsFeedStoreSpecs {
-    func test_retrieve_deliversEmptyOnEmptyCache()
-    func test_retrieve_hasNoSideEffects()
-    func test_retrieve_deliversFoundValuesOnNonEmptyCache()
-    func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
-    func test_insert_overridesPreviouslyInsertedCacheValues()
-    func test_delete_hasNoSideEffectsOnEmptyCache()
-    func test_delete_emptiesPreviousInsertedCache()
-    func test_storeSideEffects_runsSerially()
-}
+typealias FailableHackrNewsFeedStore = HackrNewsFeedStoreSpecs & FailableRetrieveHackrNewsFeedStoreSpecs &
+    FailableInsertHackrNewsFeedStoreSpecs & FailableDeleteHackrNewsFeedStoreSpecs
 
-protocol FailableRetrieveHackrNewsFeedStoreSpecs {
-    func test_retrieve_deliversErrorOnRetrievalError()
-    func test_retrieve_hasNoSideEffectsOnFailure()
-}
-
-protocol FailableInsertHackrNewsFeedStoreSpecs {
-    func test_insert_deliversErrorOnInsertionError()
-    func test_insert_hasNoSideEffectsOnInsertionError()
-}
-
-protocol FailableDeleteHackrNewsFeedStoreSpecs {
-    func test_delete_deliversErrorOnDeletionError()
-    func test_delete_hasNoSideEffectsOnDeletionError()
-}
-
-final class CodableHackrNewsFeedStoreTests: XCTestCase {
+final class CodableHackrNewsFeedStoreTests: XCTestCase, FailableHackrNewsFeedStore {
     override func setUp() {
         super.setUp()
         setupEmptyStoreState()
