+++ HackrNewsTests/Stories Cache/HackrNewsFeedStoreSpecs.swift
@@ -1,0 +1,31 @@
+//
+//  Copyright © 2021 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+protocol HackrNewsFeedStoreSpecs {
+    func test_retrieve_deliversEmptyOnEmptyCache()
+    func test_retrieve_hasNoSideEffects()
+    func test_retrieve_deliversFoundValuesOnNonEmptyCache()
+    func test_retrieve_hasNoSideEffectsOnNonEmptyCache()
+    func test_insert_overridesPreviouslyInsertedCacheValues()
+    func test_delete_hasNoSideEffectsOnEmptyCache()
+    func test_delete_emptiesPreviousInsertedCache()
+    func test_storeSideEffects_runsSerially()
+}
+
+protocol FailableRetrieveHackrNewsFeedStoreSpecs: HackrNewsFeedStoreSpecs {
+    func test_retrieve_deliversErrorOnRetrievalError()
+    func test_retrieve_hasNoSideEffectsOnFailure()
+}
+
+protocol FailableInsertHackrNewsFeedStoreSpecs: HackrNewsFeedStoreSpecs {
+    func test_insert_deliversErrorOnInsertionError()
+    func test_insert_hasNoSideEffectsOnInsertionError()
+}
+
+protocol FailableDeleteHackrNewsFeedStoreSpecs: HackrNewsFeedStoreSpecs {
+    func test_delete_deliversErrorOnDeletionError()
+    func test_delete_hasNoSideEffectsOnDeletionError()
+}
